Moved Effect execution
It's in its own file now, instead of inside each effect subclass. I think this is a better way to structure things, keeping the data with the data and putting the implementation elsewhere, kinda like a header file. I'm not super sure about the best way to lay it all out though.